requirements.txt
Flask
Flask-Discord
Flask-Login
Flask-WTF
SQLAlchemy
Flask-Migrate
python-dotenv
Flask-CORS